Use the TokenSource interface for the client credentials exchange
